requirements.txt
yfinance
pandas 
numpy
polygon-api-client
python-dotenv
langchain_openai
plotly
dash